test(schema): organize JsonWriter coverage
